personaai-backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.4
pydantic-settings==2.10.1
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
email-validator==2.2.0
python-multipart==0.0.20
redis==6.4.0
celery==5.5.3
pytest==8.4.1
httpx==0.28.1
psycopg2-binary==2.9.10
cryptography==42.0.8
openai==1.35.13
pinecone-client==3.2.2
